requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebooks (optional)
# jupyter>=1.0.0
# ipykernel>=6.22.0

# Dataset Download
requests>=2.28.0
tqdm>=4.65.0

# Model Utilities
timm>=0.9.0  # PyTorch Image Models (pretrained models)

# Audio Processing
scipy>=1.10.0

# YouTube Scraping (bark detector pipeline)
yt-dlp>=2023.10.0
imageio-ffmpeg>=0.5.0  # Bundles ffmpeg binary for audio conversion

# Audio Playback (optional, for manual validation)
# sounddevice>=0.4.6

# Experiment Tracking (optional)
# wandb>=0.15.0
# tensorboard>=2.12.0
